extract iframes controller
